requirements.txt
binance-futures-connector
pandas
numpy
sqlite3